All Tests pass now
Had to fix a test for that. Why?

The future_single_events_cleanup should test if events are removed that
don't fit the rule.
But if we don't change the rule (i.e. remove it) then all single_events
match the rule and
the count won't change. So I added the remove_recurrence_rule to the
test. Now the test
passes.