Added a simple example for statistics on the dashboard #102
I will use this as a proposal for statistics in AA, so we
can maybe simplify this procedure in the future.
I will also add other graph types as needed, but I will wait
for @boydroid's feedback on the statistics.
The chosen statistic will be changed as soon as we have a
tag-hierarchy to only feature "top-tags".