We are humans
Added a humans.txt on root according to
http://humanstxt.org
Because we are humans and we're proud of
what we created.

All people that have a team name will
be in that list

There is also a html version
/humans
path with a html version of this file
as a team page. It is of course not
designed yet.