Shift logic from SingleEvent to new Calendar class
Made the corresponding test runnable in isolation so the
execution time is a little bit faster.

Still had to include lots of rails specific stuff so there is
not so much benefit. But since we are not touching the database
it's kaaaay.

To use this feature simply run `ruby test/unit/calendar_test.rb`