Separated ActiveAdmin from normal abilities.
Attention: This is not done yet and buggy. I 
need help to fix this and get it work.

Current problem: 

The ability to edit an event
that is assigned to the user is defined in #21,
but we also have to add the ability to edit
an event if it is in a region the user
is able to edit (line #26)

Currently I only can do one and not the other. If
I try to mix both abilities, it does not work. Sadly
the doc from cancan says this should work.