Move Vagrantfile to hacken.in project
This change will make it easier to bootstrap & test a local hacken.in
instance, without bothering with MySQL / PostgreSQL setup or even
installing the native gems. The only steps left in a fresh checkout are:

```bash
  $ bundle
  $ bundle exec rake vagrant:setup
  $ vagrant up
```

After ~ 10 minutes the machine is up and running and features the
webfrontend at *hacken.local* and both the MySQL and Postgresql
servers accessible via a local port forward to 5432 and 3306.

This will obsolete the hckingbox-project.