Fix Seeds.rb: Reset PK Sequence after adding data
This fixes a bug where you could not add any events after using
`rake db:seed` as you would run into a uniqueness constraint on the
primary key.

This was due to our seeds data setting the primary key by hand which
does not influence the primary key sequence.

As we need to set the keys manually (because we use the keys as foreign
keys again in other data sets), we need to reset the primary key
sequence for all tables.